requirements-ubuntu.txt
acme==0.31.0
asn1crypto==1.2.0
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.6.0
beautifulsoup4==4.8.1
blinker==1.4
boto==2.49.0
boto3==1.10.14
botocore==1.13.14
bs4==0.0.1
celery==4.3.0
certbot==0.31.0
certbot-nginx==0.31.0
Django==2.2.7
django-celery-beat==1.5.0
django-celery-results==1.1.2
djangorestframework==3.10.3
docutils==0.15.2
feedfinder2==0.0.4
feedparser==5.2.1
future==0.15.2
gensim==3.7.1
gspread==3.1.0
gunicorn==19.7.1
httplib2==0.14.0
hyperlink==17.3.1
idna==2.8
importlib-metadata==0.22
incremental==16.10.1
jieba3k==0.35.1
jmespath==0.9.4
joblib==0.14.0
josepy==1.1.0
JPype1==0.7.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
konlpy==0.5.1
lxml==4.4.1
Markdown==3.1.1
MarkupSafe==1.0
mock==2.0.0
more-itertools==7.2.0
ndg-httpsclient==0.4.4
netifaces==0.10.4
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.17.4
oauth2client==4.1.3
oauthlib==2.0.6
packaging==19.1
pandas==0.25.1
parsedatetime==2.4
pbr==3.1.1
Pillow==6.2.1
pluggy==0.13.0
psutil==5.6.5
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyRFC3339==1.0
pyserial==3.4
pytest==5.1.2
pytest-django==3.5.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-file==1.4.3
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.2
SecretStorage==2.3.1
six==1.13.0
sklearn==0.0
smart-open==1.9.0
soupsieve==1.9.5
soynlp==0.0.493
sqlparse==0.3.0
tinysegmenter==0.3
tldextract==2.2.2
urllib3==1.25.7
wcwidth==0.1.7
xlrd==1.2.0
zipp==0.6.0